+++ src/main/java/co/uptc/frw/proyectomongo/service/CharacteristicsService.java
@@ -14,6 +14,9 @@ public class CharacteristicsService {
    @Autowired
     private CharacteristicsRepository characteristicsRepository;
 
+    @Autowired
+    private AuditService auditService;
+
     public List<CharacteristicsV> findAll() {
         return characteristicsRepository.findAll();
     }
@@ -23,11 +26,18 @@ public class CharacteristicsService {
     }
 
     public CharacteristicsV save(CharacteristicsV characteristicsV) {
+
+        CharacteristicsV savedCharacteristics = characteristicsRepository.save(characteristicsV);
+        auditService.logAudit("CharacteristicsV", "Creado: " + savedCharacteristics.toString());
         return characteristicsRepository.save(characteristicsV);
     }
 
     public void deleteCharacteristics(long id) {
-        characteristicsRepository.deleteById(id);
+        CharacteristicsV characteristicsToDelete = findById(id);
+        if (characteristicsToDelete != null) {
+            characteristicsRepository.deleteById(id);
+            auditService.logAudit("CharacteristicsV", "Borrado: ID " + id + ", Detalles: " + characteristicsToDelete.toString());
+        }
     }
 
     public CharacteristicsV update(CharacteristicsV newCharacteristic,long id) {
@@ -36,6 +46,8 @@ public class CharacteristicsService {
             characteristicsID.setBrand(newCharacteristic.getBrand());
             characteristicsID.setModel (newCharacteristic.getModel());
             characteristicsID.setLine (newCharacteristic.getLine());
+            CharacteristicsV updatedCharacteristics = save(characteristicsID);
+            auditService.logAudit("CharacteristicsV", "Modificado: " + updatedCharacteristics.toString());
             return save(characteristicsID);
         }
         throw new RuntimeException("Characteristics not found");
